Move mws python implementations to their own file
